use signed int8_ts for rack counts to fix inferences

+++ src/ent/bit_rack.h
@@ -99,7 +99,7 @@ static inline BitRack bit_rack_create_from_rack(const LetterDistribution *ld,
   BitRack bit_rack = {0, 0};
 #endif
   for (int ml = 0; ml < ld->size; ml++) {
-    const int num_this = rack_get_letter(rack, ml);
+    const int8_t num_this = rack_get_letter(rack, ml);
 #if USE_INT128_INTRINSIC
     bit_rack |= (unsigned __int128)num_this << (ml * BIT_RACK_BITS_PER_LETTER);
 #else
